+++ lib/coinbase/client/models/build_staking_operation_request.rb
@@ -14,61 +14,29 @@ require 'date'
 require 'time'
 
 module Coinbase::Client
-  # An onchain transaction.
-  class Transaction
+  class BuildStakingOperationRequest
     # The ID of the blockchain network
     attr_accessor :network_id
 
-    # The onchain address of the sender
-    attr_accessor :from_address_id
+    # The ID of the asset being staked
+    attr_accessor :asset_id
 
-    # The unsigned payload of the transaction. This is the payload that needs to be signed by the sender.
-    attr_accessor :unsigned_payload
+    # The onchain address from which the staking transaction originates and is responsible for signing the transaction.
+    attr_accessor :address_id
 
-    # The signed payload of the transaction. This is the payload that has been signed by the sender.
-    attr_accessor :signed_payload
+    # The type of staking operation
+    attr_accessor :action
 
-    # The hash of the transaction
-    attr_accessor :transaction_hash
-
-    # The link to view the transaction on a block explorer. This is optional and may not be present for all transactions.
-    attr_accessor :transaction_link
-
-    # The status of the transaction
-    attr_accessor :status
-
-    class EnumAttributeValidator
-      attr_reader :datatype
-      attr_reader :allowable_values
-
-      def initialize(datatype, allowable_values)
-        @allowable_values = allowable_values.map do |value|
-          case datatype.to_s
-          when /Integer/i
-            value.to_i
-          when /Float/i
-            value.to_f
-          else
-            value
-          end
-        end
-      end
-
-      def valid?(value)
-        !value || allowable_values.include?(value)
-      end
-    end
+    attr_accessor :options
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
         :'network_id' => :'network_id',
-        :'from_address_id' => :'from_address_id',
-        :'unsigned_payload' => :'unsigned_payload',
-        :'signed_payload' => :'signed_payload',
-        :'transaction_hash' => :'transaction_hash',
-        :'transaction_link' => :'transaction_link',
-        :'status' => :'status'
+        :'asset_id' => :'asset_id',
+        :'address_id' => :'address_id',
+        :'action' => :'action',
+        :'options' => :'options'
       }
     end
 
@@ -81,12 +49,10 @@ module Coinbase::Client
     def self.openapi_types
       {
         :'network_id' => :'String',
-        :'from_address_id' => :'String',
-        :'unsigned_payload' => :'String',
-        :'signed_payload' => :'String',
-        :'transaction_hash' => :'String',
-        :'transaction_link' => :'String',
-        :'status' => :'String'
+        :'asset_id' => :'String',
+        :'address_id' => :'String',
+        :'action' => :'String',
+        :'options' => :'Hash<String, String>'
       }
     end
 
@@ -100,13 +66,13 @@ module Coinbase::Client
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Coinbase::Client::Transaction` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Coinbase::Client::BuildStakingOperationRequest` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Coinbase::Client::Transaction`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Coinbase::Client::BuildStakingOperationRequest`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
@@ -117,34 +83,30 @@ module Coinbase::Client
         self.network_id = nil
       end
 
-      if attributes.key?(:'from_address_id')
-        self.from_address_id = attributes[:'from_address_id']
+      if attributes.key?(:'asset_id')
+        self.asset_id = attributes[:'asset_id']
       else
-        self.from_address_id = nil
+        self.asset_id = nil
       end
 
-      if attributes.key?(:'unsigned_payload')
-        self.unsigned_payload = attributes[:'unsigned_payload']
+      if attributes.key?(:'address_id')
+        self.address_id = attributes[:'address_id']
       else
-        self.unsigned_payload = nil
+        self.address_id = nil
       end
 
-      if attributes.key?(:'signed_payload')
-        self.signed_payload = attributes[:'signed_payload']
-      end
-
-      if attributes.key?(:'transaction_hash')
-        self.transaction_hash = attributes[:'transaction_hash']
-      end
-
-      if attributes.key?(:'transaction_link')
-        self.transaction_link = attributes[:'transaction_link']
-      end
-
-      if attributes.key?(:'status')
-        self.status = attributes[:'status']
+      if attributes.key?(:'action')
+        self.action = attributes[:'action']
       else
-        self.status = nil
+        self.action = nil
+      end
+
+      if attributes.key?(:'options')
+        if (value = attributes[:'options']).is_a?(Hash)
+          self.options = value
+        end
+      else
+        self.options = nil
       end
     end
 
@@ -157,16 +119,20 @@ module Coinbase::Client
         invalid_properties.push('invalid value for "network_id", network_id cannot be nil.')
       end
 
-      if @from_address_id.nil?
-        invalid_properties.push('invalid value for "from_address_id", from_address_id cannot be nil.')
+      if @asset_id.nil?
+        invalid_properties.push('invalid value for "asset_id", asset_id cannot be nil.')
       end
 
-      if @unsigned_payload.nil?
-        invalid_properties.push('invalid value for "unsigned_payload", unsigned_payload cannot be nil.')
+      if @address_id.nil?
+        invalid_properties.push('invalid value for "address_id", address_id cannot be nil.')
       end
 
-      if @status.nil?
-        invalid_properties.push('invalid value for "status", status cannot be nil.')
+      if @action.nil?
+        invalid_properties.push('invalid value for "action", action cannot be nil.')
+      end
+
+      if @options.nil?
+        invalid_properties.push('invalid value for "options", options cannot be nil.')
       end
 
       invalid_properties
@@ -177,22 +143,11 @@ module Coinbase::Client
     def valid?
       warn '[DEPRECATED] the `valid?` method is obsolete'
       return false if @network_id.nil?
-      return false if @from_address_id.nil?
-      return false if @unsigned_payload.nil?
-      return false if @status.nil?
-      status_validator = EnumAttributeValidator.new('String', ["pending", "broadcast", "complete", "failed"])
-      return false unless status_validator.valid?(@status)
+      return false if @asset_id.nil?
+      return false if @address_id.nil?
+      return false if @action.nil?
+      return false if @options.nil?
       true
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] status Object to be assigned
-    def status=(status)
-      validator = EnumAttributeValidator.new('String', ["pending", "broadcast", "complete", "failed"])
-      unless validator.valid?(status)
-        fail ArgumentError, "invalid value for \"status\", must be one of #{validator.allowable_values}."
-      end
-      @status = status
     end
 
     # Checks equality by comparing each attribute.
@@ -201,12 +156,10 @@ module Coinbase::Client
       return true if self.equal?(o)
       self.class == o.class &&
           network_id == o.network_id &&
-          from_address_id == o.from_address_id &&
-          unsigned_payload == o.unsigned_payload &&
-          signed_payload == o.signed_payload &&
-          transaction_hash == o.transaction_hash &&
-          transaction_link == o.transaction_link &&
-          status == o.status
+          asset_id == o.asset_id &&
+          address_id == o.address_id &&
+          action == o.action &&
+          options == o.options
     end
 
     # @see the `==` method
@@ -218,7 +171,7 @@ module Coinbase::Client
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [network_id, from_address_id, unsigned_payload, signed_payload, transaction_hash, transaction_link, status].hash
+      [network_id, asset_id, address_id, action, options].hash
     end
 
     # Builds the object from hash